refactor(tw): ResolvedDisplay moves beside resolveDisplay, story page follows the component name
