+++ apps/tailwind-components/app/types/display.ts
@@ -1,5 +1,3 @@
-import type { IColumn } from "../../../metadata-utils/src/types";
-
 export const LAYOUTS = ["TABLE", "CARDS", "LIST", "LINKS", "BULLETS"] as const;
 export type Layout = (typeof LAYOUTS)[number];
 
@@ -11,14 +9,4 @@ export interface DisplayConfig {
   descriptionColumnId?: string;
   detailColumnIds?: string[]; // rendered in this order
   logoColumnId?: string;
-}
-
-// What every layout component receives: slots filled in, columns resolved.
-export interface ResolvedDisplay {
-  layout: Layout;
-  titleTemplate: string;
-  subtitleTemplate?: string;
-  descriptionColumn?: IColumn;
-  detailColumns: IColumn[];
-  logoColumn?: IColumn;
 }
